refactor: separate testkit core from scenarios
